+++ src/main/java/lotto/domain/lotto/LottoNumber.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import java.util.Objects;
 import lotto.utils.StringValidator;
