+++ src/main/java/lotto/domain/lotto/LottoTicket.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import java.util.ArrayList;
 import java.util.Arrays;
