+++ src/main/java/lotto/domain/lotto/WinningNumbers.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 public class WinningNumbers {
