+++ src/main/java/lotto/domain/result/EarningRate.java
@@ -1,10 +1,10 @@
-package lotto.domain;
+package lotto.domain.result;
 
 import java.math.BigInteger;
 
 public class EarningRate {
 
-    private final BigInteger earningRate; // TODO 수익률 객체로 포장?
+    private final BigInteger earningRate;
 
     public EarningRate(BigInteger totalPrize, BigInteger buyPrice) {
         this.earningRate = (totalPrize.multiply(BigInteger.valueOf(100)).divide(buyPrice));
