+++ src/main/java/lotto/domain/result/LottoMatcher.java
@@ -1,0 +1,5 @@
+package lotto.domain.result;
+
+public class LottoMatcher {
+
+}
