--- src/main/resources/templates/index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>로또</title>
-</head>
-<body>
-Hello World!!
-</body>
-</html>
