+++ src/test/java/lotto/domain/lotto/LottoTicketTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
