+++ src/test/java/lotto/domain/result/EarningRateTest.java
@@ -1,0 +1,5 @@
+package lotto.domain.result;
+
+class EarningRateTest {
+
+}
