+++ src/test/java/lotto/domain/result/LottoMatcherTest.java
@@ -1,0 +1,5 @@
+package lotto.domain.result;
+
+class LottoMatcherTest {
+
+}
